bot/requirements.txt
git+https://github.com/Pycord-Development/pycord.git@refs/pull/2651/head
pyyaml